# api_specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Solar-Eye API",
    "version": "1.0.0",
    "description": "AI 기반 태양광 패널 결함 탐지 및 유지보수 솔루션 API 명세서"
  },
  "servers": [
    {
      "url": "https://api.solar-eye.com",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development Server"
    }
  ],
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "소셜 로그인 후 사용자 등록/조회",
        "description": "Firebase ID Token을 검증하고 사용자 정보를 등록 또는 조회합니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["firebase_token"],
                "properties": {
                  "firebase_token": {
                    "type": "string",
                    "description": "Firebase ID Token",
                    "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                  },
                  "fcm_token": {
                    "type": "string",
                    "description": "FCM 푸시 알림 토큰 (선택)",
                    "example": "fMx7kP3..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 1 },
                        "firebase_uid": { "type": "string", "example": "abc123xyz" },
                        "email": { "type": "string", "example": "user@example.com" },
                        "display_name": { "type": "string", "example": "홍길동" },
                        "provider": { "type": "string", "enum": ["google.com", "kakao.com"], "example": "google.com" },
                        "created_at": { "type": "string", "format": "date-time", "example": "2026-01-15T10:30:00Z" },
                        "last_login_at": { "type": "string", "format": "date-time", "example": "2026-01-16T14:00:00Z" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "현재 사용자 정보 조회",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "사용자 정보 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 1 },
                        "firebase_uid": { "type": "string", "example": "abc123xyz" },
                        "email": { "type": "string", "example": "user@example.com" },
                        "display_name": { "type": "string", "example": "홍길동" },
                        "provider": { "type": "string", "example": "google.com" },
                        "profile_image_url": { "type": "string", "example": "https://..." },
                        "notification_enabled": { "type": "boolean", "example": true },
                        "created_at": { "type": "string", "format": "date-time" },
                        "last_login_at": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/fcm-token": {
      "post": {
        "tags": ["Authentication"],
        "summary": "FCM 토큰 등록/갱신",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fcm_token"],
                "properties": {
                  "fcm_token": {
                    "type": "string",
                    "description": "Firebase Cloud Messaging 토큰",
                    "example": "fMx7kP3R8..."
                  },
                  "device_type": {
                    "type": "string",
                    "enum": ["android", "ios"],
                    "example": "android"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "토큰 등록 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "FCM 토큰이 등록되었습니다." }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/detections": {
      "get": {
        "tags": ["Detections"],
        "summary": "탐지 이력 조회",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "panel_id",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "패널 ID 필터"
          },
          {
            "name": "defect_type",
            "in": "query",
            "schema": { "type": "string", "enum": ["defect", "soiling", "normal"] },
            "description": "결함 유형 필터"
          },
          {
            "name": "start_date",
            "in": "query",
            "schema": { "type": "string", "format": "date" },
            "description": "조회 시작일 (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": { "type": "string", "format": "date" },
            "description": "조회 종료일 (YYYY-MM-DD)"
          },
          {
            "name": "min_confidence",
            "in": "query",
            "schema": { "type": "number", "minimum": 0, "maximum": 1 },
            "description": "최소 신뢰도 필터"
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "탐지 이력 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Detection" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/detections/{detection_id}": {
      "get": {
        "tags": ["Detections"],
        "summary": "탐지 상세 조회",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "detection_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "탐지 상세 정보",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/DetectionDetail" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID Token",
        "description": "Firebase Authentication ID Token"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "UNAUTHORIZED" },
              "message": { "type": "string", "example": "인증 토큰이 유효하지 않습니다." }
            }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 20 },
          "total_items": { "type": "integer", "example": 100 },
          "total_pages": { "type": "integer", "example": 5 },
          "has_next": { "type": "boolean", "example": true },
          "has_prev": { "type": "boolean", "example": false }
        }
      },
      "Panel": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "서울 태양광 발전소 A동" },
          "location": { "type": "string", "example": "서울시 강남구 역삼동" },
          "rtsp_url": { "type": "string", "example": "rtsp://..." },
          "status": { "type": "string", "enum": ["active", "inactive", "error"], "example": "active" },
          "latitude": { "type": "number", "example": 37.5665 },
          "longitude": { "type": "number", "example": 126.9780 },
          "grid_nx": { "type": "integer", "example": 60 },
          "grid_ny": { "type": "integer", "example": 127 },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Detection": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "panel_id": { "type": "integer", "example": 1 },
          "defect_type": { "type": "string", "enum": ["defect", "soiling", "normal"], "example": "soiling" },
          "defect_subtype": { "type": "string", "example": "bird_drop", "description": "상세 결함 유형 (dust, bird_drop, crack 등)" },
          "confidence": { "type": "number", "format": "float", "example": 0.92 },
          "bbox": {
            "type": "object",
            "properties": {
              "x": { "type": "number", "example": 100.5 },
              "y": { "type": "number", "example": 200.3 },
              "width": { "type": "number", "example": 50.0 },
              "height": { "type": "number", "example": 60.0 }
            },
            "description": "바운딩 박스 좌표"
          },
          "snapshot_url": { "type": "string", "example": "https://storage.googleapis.com/solar-eye/snapshots/2026/01/16/abc123.jpg" },
          "detected_at": { "type": "string", "format": "date-time", "example": "2026-01-16T14:30:00Z" }
        }
      },
      "DetectionDetail": {
        "type": "object",
        "allOf": [
          { "$ref": "#/components/schemas/Detection" },
          {
            "type": "object",
            "properties": {
              "panel": { "$ref": "#/components/schemas/Panel" },
              "weather_at_detection": { "$ref": "#/components/schemas/CurrentWeather" },
              "alert": { "$ref": "#/components/schemas/Alert" }
            }
          }
        ]
      },
      "Alert": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "detection_id": { "type": "integer", "example": 1 },
          "user_id": { "type": "integer", "example": 1 },
          "alert_type": { "type": "string", "enum": ["defect", "soiling", "system"], "example": "defect" },
          "title": { "type": "string", "example": "🔴 패널 결함 감지" },
          "message": { "type": "string", "example": "서울 태양광 발전소 A동에서 패널 파손이 감지되었습니다." },
          "is_read": { "type": "boolean", "example": false },
          "snapshot_url": { "type": "string", "example": "https://..." },
          "sent_at": { "type": "string", "format": "date-time" },
          "read_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
      "CurrentWeather": {
        "type": "object",
        "properties": {
          "location": {
            "type": "object",
            "properties": {
              "nx": { "type": "integer", "example": 60, "description": "기상청 격자 X 좌표" },
              "ny": { "type": "integer", "example": 127, "description": "기상청 격자 Y 좌표" }
            }
          },
          "temperature": { "type": "number", "example": 5.2, "description": "현재 기온 (°C)" },
          "humidity": { "type": "integer", "example": 45, "description": "습도 (%)" },
          "precipitation_type": { 
            "type": "string", 
            "enum": ["none", "rain", "rain_snow", "snow", "shower"],
            "example": "none",
            "description": "강수형태 (PTY: 0=없음, 1=비, 2=비/눈, 3=눈, 4=소나기)"
          },
          "sky_condition": {
            "type": "string",
            "enum": ["clear", "partly_cloudy", "cloudy", "overcast"],
            "example": "clear",
            "description": "하늘 상태 (SKY: 1=맑음, 3=구름많음, 4=흐림)"
          },
          "wind_speed": { "type": "number", "example": 2.5, "description": "풍속 (m/s)" },
          "precipitation_1h": { "type": "number", "example": 0, "description": "1시간 강수량 (mm)" },
          "is_rainy": { "type": "boolean", "example": false, "description": "우천 여부" },
          "condition_text": { "type": "string", "example": "맑음", "description": "날씨 상태 텍스트" },
          "condition_icon": { "type": "string", "example": "sunny", "description": "날씨 아이콘 코드" },
          "observed_at": { "type": "string", "format": "date-time", "example": "2026-01-16T14:00:00Z" },
          "fetched_at": { "type": "string", "format": "date-time", "example": "2026-01-16T14:05:00Z" }
        }
      },
...
    }
  }
}